requirements.txt
pydantic
protobuf
paho-mqtt
sparkplug_b_packets
redis[hiredis]
asyncer